# docs/swagger.yaml
basePath: /
definitions:
  helpers.Response:
    properties:
      detail:
        type: string
      request:
        type: string
      time:
        type: string
      title:
        type: string
    type: object
  helpers.Verse:
    properties:
      verse:
        example: |-
          Ooh baby, don't you know I suffer?
          Ooh baby, can you hear me moan?
          You caught me under false pretenses
          How long before you let me go?
        type: string
    type: object
  model.Song:
    properties:
      group:
        example: Muse
        type: string
      id:
        type: string
      link:
        default: NOT FOUND
        example: https://www.youtube.com/watch?v=Xsp3_a-PMTw
        type: string
      releaseDate:
        default: NOT FOUND
        example: 16.07.2006
        type: string
      song:
        example: Supermassive Black Hole
        type: string
      text:
        default: NOT FOUND
        example: |-
          Ooh baby, don't you know I suffer?
          Ooh baby, can you hear me moan?
          You caught me under false pretenses
          How long before you let me go?

          Ooh
          You set my soul alight
          Ooh
          You set my soul alight
        type: string
    type: object
  model.SongCreate:
    properties:
      group:
        example: Muse
        type: string
      song:
        example: Supermassive Black Hole
        type: string
    required:
    - group
    - song
    type: object
  model.SongUpdate:
    properties:
      group:
        example: Muse
        type: string
      link:
        example: https://www.youtube.com/watch?v=Xsp3_a-PMTw
        type: string
      releaseDate:
        example: 16.07.2006
        type: string
      song:
        example: Supermassive Black Hole
        type: string
      text:
        example: |-
          Ooh baby, don't you know I suffer?
          Ooh baby, can you hear me moan?
          You caught me under false pretenses
          How long before you let me go?

          Ooh
          You set my soul alight
          Ooh
          You set my soul alight
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Server for Music library application
  title: Music library API
  version: "1.0"
paths:
  /v1/all:
    get:
      description: Посмотреть все доступные песни с данными
      operationId: get-all
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Song'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      summary: Get All from mockserver
      tags:
      - MockServer
  /v1/songs:
    get:
      consumes:
      - application/json
      description: get songs from library
      operationId: get-song
      parameters:
      - description: Filter by group
        in: query
        name: group
        type: string
      - description: Filter by song
        in: query
        name: song
        type: string
      - description: Filter by created_at
        in: query
        name: releaseDate
        type: string
      - description: Filter by text
        in: query
        name: text
        type: string
      - description: Filter by link
        in: query
        name: link
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Song'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      summary: Get songs
      tags:
      - Song
    post:
      consumes:
      - application/json
      description: add song to library
      operationId: create-song
      parameters:
      - description: song info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.SongCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      summary: Create song
      tags:
      - Song
  /v1/songs/{id}:
    delete:
      consumes:
      - application/json
      description: delete song from library
      operationId: delete-song
      parameters:
      - description: delete by id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      summary: Delete song
      tags:
      - Song
    put:
      consumes:
      - application/json
      description: Update song
      operationId: update-song
      parameters:
      - description: update by id
        in: path
        name: id
        required: true
        type: string
      - description: song info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.SongUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      summary: Update song
      tags:
      - Song
  /v1/songs/verse/{id}:
    get:
      consumes:
      - application/json
      description: get verse of song
      operationId: get-verse
      parameters:
      - description: Song id
        in: path
        name: id
        type: string
      - description: Number of verse (номер куплета)
        in: query
        name: num
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Verse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      summary: Get verse
      tags:
      - Song
swagger: "2.0"